Move addDocumentAndData to documents package.
